Switch to using the URL ID as a key
Previously, this scraper was using the name as a key - this isn't a good
key, since there might be multiple people with the same name, or even if
there aren't, minor changes in the name (e.g. adding or removing middle
names, encoding problems) will cause a new person to be created.
Tony Bowden (@tmtmtmtm) suggested using the ID in the person page's URL
instead - this commit makes that change.